src-api/requirements.txt
argon2-cffi==20.1.0
Django==3.0.6
django-cors-headers==3.3.0
django-mysql==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
mysqlclient==1.4.6
pycodestyle==2.6.0
pylint==2.5.2
pytest==5.4.2
